refactor policy controller
